% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{bold_significant}
\alias{bold_significant}
\title{Bold Cells with Significant P-values}
\usage{
bold_significant(
  ft,
  col_index,
  column,
  skip_rows = NULL,
  alpha = "05",
  decimal_place = 6
)
}
\arguments{
\item{ft}{A flextable object.}

\item{col_index}{The column index for the column in the
flextable object to update.}

\item{column}{The character vector for the column of p-values
(by default, assumes p-values are in the format 'p = 0.X',
where the decimal place is the 6th character in the string).}

\item{skip_rows}{An optional logical vector of matching length
to \code{column} marked \code{TRUE} for rows to skip
when processing.}

\item{alpha}{A character string with the cut-off for significance.
For example, if p < 0.05 is deemed significant, \code{alpha}
would be '05'. If p < 0.005 is deemed significant, then
\code{alpha} would be '005'.}

\item{decimal_place}{The index for when the decimal place in
the character string representing p-values starts. Values
after this index will be compared against \code{alpha} to
assess for significance, values before and at this index
will be ignored.}
}
\value{
An updated flextable object with bolded cells for
significant p-values for the specified column.
}
\description{
A function that will check whether a column in an
APA-style flextable object has significant p-values;
if it does, these cells are bolded.
}
\examples{
# Create example table
ex = data.frame(
  Test = c( 'Test 1', 'A', 'B', 'Test 2', 'A', 'B' ),
  p_value = c( '', 'p = 0.123', 'p = 0.040', '', 'p = 0.051', 'p < 0.001' )
)
th = create_header( ex ); th$colA = c( 'Label', 'p-value' )
ft = create_ft( ex, th )

# Bold cells with p < 0.05
ft.1 = ft \%>\% bold_significant( 2, ex$p_value, alpha = '05' )
ft.1

# Bold cells with p < 0.005
ft.2 = ft \%>\% bold_significant( 2, ex$p_value, alpha = '005' )
ft.2

}
