% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{create_ft}
\alias{create_ft}
\title{Function to Create an APA-style Flextable Object}
\usage{
create_ft(
  tbl,
  tbl_header,
  alignment = NULL,
  font.size = 11,
  font.family = "Arial",
  padding_sym = "   ",
  padding_pos = "left",
  alternate_col = NULL
)
}
\arguments{
\item{tbl}{A data frame.}

\item{tbl_header}{A data frame with the structure for
the header of the flextable object (see
\code{\link{create_header}}).}

\item{alignment}{An optional named list giving the
column indices to be either left, right, or
center-aligned.}

\item{font.size}{The size of the font (defaults to 11).}

\item{font.family}{The font family (defaults to Arial).}

\item{padding_sym}{The set of characters to search for
when checking for padding within cells.}

\item{padding_pos}{Whether padding should be applied to
the left or right.}

\item{alternate_col}{An optional color specification to
create alternating rows shaded a different color.}
}
\value{
A processed character string.
}
\description{
A function that converts a data frame into a
APA-style flextable object.
}
\examples{
# Example table
data( "mtcars" )
tbl = mtcars[1:6,1:2]
tbl = cbind( Cars = rownames( tbl ), tbl )
# Table header
th = create_header( tbl )
th$colA = c( "Cars", "MPG", "Cylinders" )
# List with column indices for alignments
a = list( left = 1, center = 2, right = 3 )
# Create flextable object
ft = create_ft( tbl, th, alignment = a )

}
