% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{create_header}
\alias{create_header}
\title{Function to Initialize Header for Flextable Object}
\usage{
create_header(tbl, nc = 1)
}
\arguments{
\item{tbl}{The data frame to be converted into a
flextable object.}

\item{nc}{The number of columns (later converted
into rows for the header) to initialize.}
}
\value{
A data frame with the 'col_keys' column
  and the columns corresponding to the rows to
  add to the header (with placeholder text).
}
\description{
A function to initialize a data frame with the
header information needed when creating a flextable
object.
}
\examples{
# Example table
data( "mtcars" )
tbl = mtcars[1:6,1:3]
tbl = cbind( Cars = rownames( tbl ), tbl )
# Initialize header
th = create_header( tbl, nc = 2 )
# Replace placeholder text
th$colA = c( "Cars", "Miles", "Cylinders", "Displacement" )
th$colB = c( "", "(per gallon)", "(Number)", "(cu. in.)" )
# Create flextable object
ft = create_ft( tbl, th )

}
