% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extofficer-package.R
\docType{package}
\name{extofficer-package}
\alias{extofficer-package}
\alias{extofficer}
\title{Convenience Functions for the 'officer' Package}
\description{
The \pkg{extofficer} package implements a set of
convenience functions that interface with the
packages \pkg{officer} and \pkg{flextable} to aid
in quickly creating simple Word documents with nice
tables and figures. A complete  list of the functions
available in the package can be obtained via the command
\code{ls(pos = "package:extofficer")}.
}
