% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{word_add_image}
\alias{word_add_image}
\title{Function to Add Image to a docx Object}
\usage{
word_add_image(
  x,
  filename,
  string = NULL,
  num = NULL,
  width = 6,
  height = 6,
  ...
)
}
\arguments{
\item{x}{A docx device.}

\item{filename}{The image filename.}

\item{string}{A figure caption.}

\item{num}{The figure number.}

\item{width}{The height of the image in inches.}

\item{height}{The width of the image in inches.}

\item{...}{Additional parameters for the
\code{\link{word_add_text}} function.}
}
\value{
An updated docx object.
}
\description{
A function that streamlines the process of adding
an image to a docx object via the \pkg{officer}
package.
}
\examples{
# Create docx object
my_doc = read_docx()

# Save example png
fname = "Example.png"
png( filename = fname, width = 6, height = 6, units = "in", res = 200 )
hist( rnorm(100), col = 'grey', border = 'white', breaks = 'FD' )
dev.off()

# Add to docx object
my_doc = my_doc \%>\% word_add_image( fname, string = "Example", num = 1 )

# Write to a docx file
my_doc \%>\% print( target = tempfile(fileext = ".docx" ) )
# full path of produced file is returned
print(.Last.value)

}
