% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{word_add_table}
\alias{word_add_table}
\title{Function to Add Flextable Object to a docx Object}
\usage{
word_add_table(x, ft, string = NULL, num = NULL, align = "left", ...)
}
\arguments{
\item{x}{A docx device.}

\item{ft}{A flextable object.}

\item{string}{A table caption.}

\item{num}{The table number.}

\item{...}{Additional parameters for the
\code{\link{word_add_text}} function.}
}
\value{
An updated docx object.
}
\description{
A function that streamlines the process of adding
a flextable object to a docx object via the
\pkg{officer} and \pkg{flextable} packages.
}
\examples{
# Create docx object
my_doc = read_docx()

# Example table
data( "mtcars" )
tbl = mtcars[1:6,1:2]
tbl = cbind( Cars = rownames( tbl ), tbl )
# Table header
th = create_header( tbl )
th$colA = c( "Cars", "MPG", "Cylinders" )
# List with column indices for alignments
a = list( left = 1, center = 2, right = 3 )
# Create flextable object
ft = create_ft( tbl, th, alignment = a )

# Add to docx object
my_doc = my_doc \%>\% word_add_text( ft, string = "Example", num = 1 )

# Write to a docx file
my_doc \%>\% print( target = tempfile(fileext = ".docx" ) )
# full path of produced file is returned
print(.Last.value)

}
