% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{word_add_text}
\alias{word_add_text}
\title{Function to Add Formatted Text to a docx Object}
\usage{
word_add_text(x, string, font.size = 11, font.family = "Arial", ...)
}
\arguments{
\item{x}{A docx device.}

\item{string}{A character string.}

\item{font.size}{The font size.}

\item{font.family}{The font family.}

\item{...}{Additional parameters for the
\code{\link[officer]{fp_text}} function.}
}
\value{
An updated docx object.
}
\description{
A function that streamlines the process of adding
a formatted paragraph to a docx object generated
via the \pkg{officer} package.
}
\examples{
# Create docx object
require( "officer" )
my_doc = read_docx()

# Example text
string = "Hello word!"
my_doc = my_doc \%>\% word_add_text( string )

# Write to a docx file
my_doc \%>\% print( target = tempfile(fileext = ".docx" )
# full path of produced file is returned
print(.Last.value)

}
