% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{wtext}
\alias{wtext}
\title{Function to Prepare Text to Add to a Word Document}
\usage{
wtext(string, strip_nl = F, strip_dbl = F, values = NULL)
}
\arguments{
\item{string}{A character string.}

\item{strip_nl}{Logical; if \code{TRUE}, strips the
new line indicator '\code{\\n}' from the sring.}

\item{strip_dbl}{Logical; if \code{TRUE}, strips
instances of double spaces from the string.}

\item{values}{An optional vector of values (e.g., numeric
values like descriptive statistics) to replace placeholders
of the form '\code{[[i]]}' where \code{i} refers to the
ith position in the vector \code{values}.}
}
\value{
A processed character string.
}
\description{
A function to process a character string (e.g.,
substituting values for placeholders or removing
new line characters ) before adding it to a Word
document.
}
\examples{
# Example string
string = "
Hello world!
 Here is a new sentence.
"
# Remove new line characters
wtext( string, strip_nl = T )

# Substituting values
values = c( 2, 10 )
string = "
Here is the first value: [[1]],
 and here is the second value: [[2]].
"
wtext( string, strip_nl = T, values = values )

}
